Update AD & MS365 (Cloud Conversion) -- Offboarding Script.ps1
⭐ Added Lines: 
376: # Run a Delta Sync to sync the changes to MS365 - this is so when the account becomes a cloud account, attributes in AD that were just updated are synced over.

377: Invoke-Command –ComputerName $AADSyncServer –ScriptBlock {Start-ADSyncSyncCycle –PolicyType Delta}       

388: # Pausing Script for 10 seconds. We just ran AADsync a moment ago and I want the system to catch up before AADsync is ran again.

389: start-sleep 10

⭐ Reason for Update:
I noticed that the script would not apply the changes to the offboarded account from AD to MS365. Reviewing, this was due to the fact that the account gets moved to the unsynced OU before syncing to MS365, so MS365 doesn't apply all the changes that were made to the AD account. By running a quick DeltaSync before the account is moved to the unsynced OU, it will bring over all those changes and then when AADsync is ran again, it will then soft-delete the account in MS365 and convert it to a cloud account.